Refactor `Execution.kt`
